rydde opp litt
mye vann under broa
