+++ gammelt_forsøk/mysql_settings.py
@@ -1,0 +1,3 @@
+settings = {}
+settings['user'] = 'user'
+settings['password'] = 'password'
